# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/otp/send": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Firebase otp sms send to phone number",
                "operationId": "c8f7d02983e6da901f36d42140f73a73",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "phone": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone": "+7"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "The provided credentials are incorrect."
                    }
                }
            }
        },
        "/otp/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check sms code",
                "operationId": "fc32151dccbce8173d7bdcb7e688e1ea",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "phone": {
                                                "type": "string"
                                            },
                                            "code": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone": "+7",
                                    "code": "999-999"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "The provided credentials are incorrect."
                    }
                }
            }
        },
        "/otp/firebase/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Firebase check with token",
                "operationId": "d3c41ca3090dd9f80d39a8832b00e420",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "phone": {
                                                "type": "string"
                                            },
                                            "firebase_token": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "phone": "+7",
                                    "firebase_token": "token"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "The provided credentials are incorrect."
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}